remise en double des buffers, apres melange dans les branches, une version de test avait été insérée, mais qui générait des plantages.

+++ sources/model/Cloud.cpp
@@ -795,7 +795,7 @@ Env Cloud::getEnvelopeVolume()
 
 
 // compute audio
-void Cloud::nextBuffer(float *accumBuff, unsigned int numFrames)
+void Cloud::nextBuffer(double *accumBuff, unsigned int numFrames)
 {
     int l_envelopeAction = this->envelopeAction.exchange(0);
